+++ doc/Changelog.md
@@ -3,6 +3,8 @@ Changelog {#Changelog}
 
 # git master
 
+* [172](https://github.com/BlueBrain/Brion/pull/172):
+  zeroeq::Server/Client based morphology loader
 * [171](https://github.com/BlueBrain/Brion/pull/171):
   Ban obsolete morphology repair stage from API: The stage parameter from
   certain Morphology methods has been removed. In the case of h5v2 morphologies,